Split yaml-section-values.ts into lexer/list/reader/facade layers
The file was ~1275 lines, well over the 500-600 cap, mixing low-level
lexing, block/list scanning, the recursive structural reader, and the
mutation API. Move each concern into its own strictly one-directional
layer (facade -> reader -> list -> lexer), every function body verbatim,
no behaviour change:

- yaml-section-lexer.ts  (~279) regexes + line/indent classifiers
- yaml-section-list.ts   (~184) block/list scanning primitives
- yaml-section-reader.ts (~565) parseYamlSectionValues + recursive dispatch
- yaml-section-values.ts (~312) facade: findSectionRange / updateSectionInYaml / removeSectionFromYaml

Consumers import each symbol from its owning layer (no re-export shim).
Replaces the stale, conflicting bot PR #633.
